+++ src/app/api/users/route.ts
@@ -1,4 +1,5 @@
 import { NextResponse } from 'next/server';
+
 import data from './data.json';
 
 export async function GET(request: any) {
