+++ src/app/layout.tsx
@@ -2,13 +2,14 @@
 
 import { RecoilRoot } from 'recoil';
 
-import InApp from './InApp';
+import Script from 'next/script';
+
 import './globals.css';
 import AxiosInterceptor from '@/component/Common/AxiosInterceptor';
 import Description from '@/component/Common/Description';
 import Modal from '@/component/Common/Modal';
 import OpenGraph from '@/component/Common/OpenGraph';
-import Script from 'next/script';
+import InApp from '@/component/InApp';
 
 export default function RootLayout({
   children,
