+++ src/app/login/page.tsx
@@ -1,5 +1,8 @@
 'use client';
 
+import Image from 'next/image';
+import { useRouter } from 'next/navigation';
+
 import Close from '@/component/Common/Icon/Close';
 import { bodyFontState } from '@/recoil/font/body';
 import { captionFontState } from '@/recoil/font/caption';
@@ -7,8 +10,6 @@ import { headlineFontState } from '@/recoil/font/headline';
 import { titleFontState } from '@/recoil/font/title';
 import { logging } from '@/services/mixpanel';
 import { defaultState } from '@/utils/theme/default';
-import Image from 'next/image';
-import { useRouter } from 'next/navigation';
 
 export default function Login() {
   const router = useRouter();
