+++ src/app/page.tsx
@@ -2,6 +2,9 @@
 
 import { useEffect, useRef, useState } from 'react';
 
+import Image from 'next/image';
+import { useRouter } from 'next/navigation';
+
 import HeaderLayout from '@/component/Common/HeaderLayout';
 import MainManStrcat from '@/component/MainManStrcat';
 import { useLogin } from '@/hooks/useLogin';
@@ -10,8 +13,6 @@ import { titleFontState } from '@/recoil/font/title';
 import { logging } from '@/services/mixpanel';
 import { focusToHighlight } from '@/utils/focusToHighlight';
 import { defaultState } from '@/utils/theme/default';
-import Image from 'next/image';
-import { useRouter } from 'next/navigation';
 
 export default function Home() {
   const [isLogin] = useLogin();
