+++ src/component/Common/Modal/PhotoModal.tsx
@@ -2,7 +2,7 @@ import Close from '../Icon/Close';
 import Photo from './Photo';
 import { bodyFontState } from '@/recoil/font/body';
 import { captionFontState } from '@/recoil/font/caption';
-import { themeState } from '@/recoil/newtheme/theme';
+import { themeState } from '@/recoil/theme';
 import { handleBackground } from '@/utils/handleBackground';
 
 interface Props {
