+++ src/component/Personal/BottomAnimationImage.tsx
